requirements.txt
streamlit
requests
Pillow
base64
os